fix(vite): add correct gitignore pattern for vite timestamp files #28685 (#28693)
<!-- Please make sure you have read the submission guidelines before
posting an PR -->
<!--
https://github.com/nrwl/nx/blob/master/CONTRIBUTING.md#-submitting-a-pr
-->

<!-- Please make sure that your commit message follows our format -->
<!-- Example: `fix(nx): must begin with lowercase` -->

<!-- If this is a particularly complex change or feature addition, you
can request a dedicated Nx release for this pull request branch. Mention
someone from the Nx team or the `@nrwl/nx-pipelines-reviewers` and they
will confirm if the PR warrants its own release for testing purposes,
and generate it for you if appropriate. -->

<!-- This is the behavior we have today -->
The pattern added for .gitignore is incorrect

<!-- This is the behavior we should expect with the changes in this PR
-->
Add the correct pattern

<!-- Please link the issue being fixed so it gets closed when this is
merged. -->

Fixes #28685