fix(misc)!: only provide default value for object properties if object already has value
Defaults causing an object to be defined causes some confusing behavior when a schema has defaults and required properties for a given object prop. If the required prop doesn't have a default value, and another property does, it becomes invalid to not pass the object when it would be valid had that default not been specified.

BREAKING CHANGE: Currently if an executor's schema provides some default values for an object's properties, those defaults cause the object to be defined. This changes that, such that the defaults are only applied if the object exists in the first place.