fix(core): shutdown plugin workers when disconnected
Currently we send a message to the plugin to shutdown when the parent process either dies or is ready for them to shutdown. This works, but fails when the host process is killed with a sigterm as the cleanup can't run. This PR shifts the strategy such that whenever the one and only socket connection ends, the plugin automatically shuts down