fix(core): shutdown plugin workers when disconnected (#28857)
Currently we send a message to the plugin to shutdown when the parent
process either dies or is ready for them to shutdown. This works, but
fails when the host process is killed with a sigterm as the cleanup
can't run. This PR shifts the strategy such that whenever the one and
only socket connection ends, the plugin automatically shuts down

(cherry picked from commit 0d7f2263383b2bd4fc3062a5a896a1271fd2674c)